Explicitly give username to `Dir.home`.
This is called without forking to the user, so was returning `/root` rather
than `/home/username/`.

When running under sudo `Dir.home` was correctly returning the home directory
expected (e.g. that of the user running sudo) but when run as root it meant
that user targets were being ignored.